requirements.txt
Django~=4.1.2
djangorestframework~=3.14.0
Pillow~=9.2.0
django-filter
pytest-django~=4.5.2
coverage~=6.5.0
